putting firebase into separate folder
