+++ src/components/Attendance/Attendance.vue
@@ -16,7 +16,7 @@
 <script>
   import Vue from "vue";
   import StudentLessons from "./StudentLessons.vue"
-  import { getLessons } from "../../firebase.js"
+  import { getLessons } from "@/firebase/firebase.js"
 
   export default Vue.extend({
     name: "Attendance",
