+++ src/components/Groups/Groups.js
@@ -1,7 +1,7 @@
 import Vue from "vue";
 import Teachers from "../Teachers/Teachers.vue";
 import Students from "./Students/Students.vue";
-import { createGroup, getGroups, getAllRegistrations } from "../../firebase.js";
+import { createGroup, getGroups, getAllRegistrations } from "@/firebase/firebase.js";
 
 export default Vue.extend ({
   name: "Groups",
