+++ src/components/Header/Header.js
@@ -1,7 +1,7 @@
 import Vue from "vue";
 import firebase from "firebase/app";
 import 'firebase/auth';
-import { checkUserClaim } from "../../firebase.js";
+import { checkUserClaim } from "@/firebase/firebase.js";
 import { EventBus } from "../../EventBus";
 
 export default Vue.extend({
