+++ src/main.js
@@ -5,12 +5,12 @@ import BootstrapVue from 'bootstrap-vue';
 import App from './App.vue';
 import { router } from './router';
 
-Vue.config.productionTip = false
+Vue.config.productionTip = false;
 
 Vue.use(BootstrapVue);
 
-import 'bootstrap/dist/css/bootstrap.css'
-import 'bootstrap-vue/dist/bootstrap-vue.css'
+import 'bootstrap/dist/css/bootstrap.css';
+import 'bootstrap-vue/dist/bootstrap-vue.css';
 
 let app = '';
