+++ src/router.js
@@ -8,7 +8,7 @@ import Login from './views/Login.vue';
 import SignUp from './views/SignUp.vue';
 import School from './views/School.vue';
 import ClassRoom from './views/ClassRoom.vue';
-import { checkUserClaim } from './firebase';
+import { checkUserClaim } from './firebase/firebase';
 
 Vue.use(Router);
